requirements.txt
streamlit
pillow
pandas
openai
scikit-learn